requirements.txt
chromadb>=1.0.11
gradio==5.44.1
numpy>=2.2.6
pillow>=11.2.1
sentencepiece>=0.2.0
torch
transformers>=4.57.3